// src/components/primaryBlocks/types/Blocks.ts
export enum BlockType {
  ADD_BLOCK = "ADD_BLOCK",
  SIGNUP1 = "SIGNUP1",
  FOOTER1 = "FOOTER1",
  GET_IN_TOUCH1 = "GET_IN_TOUCH1",
  HERO1 = "HERO1",
  MARYORALVEN1 = "MARYORALVEN1",
  OUR_TEAM1 = "OUR_TEAM1",
  PRICING1 = "PRICING1",
  TESTIMONIAL1 = "TESTIMONIAL1",
  FAQ1 = "FAQ1",
  HEADER2 = "HEADER2",
  ABOUT_US2 = "ABOUT_US2",
  COUNTER2 = "COUNTER2",
  VIDEO2 = "VIDEO2",
  MEETOURTEAM2 = "MEETOURTEAM2",
  ICONCLINETS2 = "ICONCLINETS2",
  WORKWITHUS2 = "WORKWITHUS2",
  HEROHEADER1 = "HEROHEADER1",
  MAYORALVENHEADER1 = "MAYORALVENHEADER1",
}

export const BlockCategories = {
  HEADER: [
    BlockType.HEADER2,
    BlockType.MAYORALVENHEADER1,
    BlockType.HEROHEADER1,
  ],
  FOOTER: [BlockType.FOOTER1],
  SIGNUP: [BlockType.SIGNUP1],
  CONTACT: [BlockType.GET_IN_TOUCH1],
  HERO: [BlockType.HERO1],
  MEGA_MENU: [BlockType.MARYORALVEN1],
  OUR_TEAM: [BlockType.OUR_TEAM1],
  PRICING: [BlockType.PRICING1],
  COUNTERS: [BlockType.COUNTER2],
  BRANDS: [BlockType.ICONCLINETS2],
  ABOUT_US: [
    BlockType.ABOUT_US2,
    BlockType.MEETOURTEAM2,
    BlockType.WORKWITHUS2,
  ],
  VIDEO: [BlockType.VIDEO2],
  FAQ: [BlockType.FAQ1],
  TESTIMONIALS: [BlockType.TESTIMONIAL1],
};